changed start of normal downlink flag
